pyproject.toml
[project]
name = "polymarket-trading-bot"
version = "2.0.0"
description = "Professional trading bot for Polymarket prediction markets"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Jose V.", email = "josev@example.com"}
]

dependencies = [
    "py-clob-client>=0.10.0",
    "python-dotenv>=1.0.0",
    "web3>=6.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
async = [
    "aiohttp>=3.9.0",
    "aiofiles>=23.0.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.6.0",
    "types-requests>=2.31.0",
]
monitoring = [
    "python-json-logger>=2.0.0",
    "prometheus-client>=0.18.0",
]

[project.scripts]
polymarket-bot = "tradingbot.main:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["tradingbot"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = [
    "--cov=tradingbot",
    "--cov-report=term-missing",
    "--cov-report=html",
    "-v"
]

[tool.black]
line-length = 100
target-version = ["py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.venv
  | __pycache__
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
]
ignore = [
    "E501",  # line too long (black handles this)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Start lenient, tighten later
ignore_missing_imports = true
files = ["tradingbot"]

[[tool.mypy.overrides]]
module = "py_clob_client.*"
ignore_missing_imports = true
